Update bootstrap-4-autocomplete.js
This code will provide the ability to do exclusive searches for each term in the search string delimited by spaces.  As an example.  if you set treshold to 2 and do bo, it will find all matches of bo.  If you then add bo ex, it will find all match that include bo and ex only.  If you then add au, it would find anything that matches bo, ex and au.  It also highlights those matches as well if that functionality is turned on.